requirements.txt
# Install with: pip3.12 install -r requirements.txt
# For PythonAnywhere: pip3.12 install --user -r requirements.txt

# Core Django packages
Django==5.1.3
djangorestframework==3.15.2

# Filtering and CORS
django-filter==24.2
django-cors-headers==4.4.0

# Database (if needed)
# psycopg2-binary==2.9.10  # Uncomment if using PostgreSQL

# Testing and HTTP requests
requests==2.31.0
